feat(desktop): single-column Cursor-style automation editor with tabs
- right rail dissolved into the centered column as Cursor-style cards
- Triggers card: sentence-style '[Daily at 8:00 AM] [tz]' with indented
  scope line 'in [project] on [device] · [workspace]' and a Next-run line
- Instructions card hosts the prompt editor with the Agent picker in the
  bottom-left slot (Cursor's model-picker position)
- Settings | Run History tabs; runs list moves to the Run History tab
  full-width with inline errors
