+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationBody/AutomationBody.tsx
@@ -1,26 +1,42 @@
-import type { SelectAutomation } from "@superset/db/schema";
+import type {
+	SelectAutomation,
+	SelectAutomationRun,
+} from "@superset/db/schema";
 import { toast } from "@superset/ui/sonner";
 import { Switch } from "@superset/ui/switch";
+import { cn } from "@superset/ui/utils";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useEffect, useRef, useState } from "react";
 import { EmojiTextInput } from "renderer/components/EmojiTextInput";
 import { MarkdownEditor } from "renderer/components/MarkdownEditor";
+import { useHostUrl } from "renderer/hooks/host-service/useHostTargetUrl";
+import { useV2AgentChoices } from "renderer/hooks/useV2AgentChoices";
 import { apiTrpcClient } from "renderer/lib/api-trpc-client";
+import { useWorkspaceHostOptions } from "renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/components/DevicePicker/hooks/useWorkspaceHostOptions/useWorkspaceHostOptions";
+import { AgentPicker } from "../../../components/AgentPicker";
 import { useProjectFileSearch } from "../../../hooks/useProjectFileSearch";
+import { matchAgentChoice } from "../../../utils/agentIdentity";
+import { PreviousRunsList } from "../PreviousRunsList";
+import { type AutomationUpdatePatch, TriggersCard } from "../TriggersCard";
+
+type DetailTab = "settings" | "runs";
 
 export function AutomationBody({
 	automation,
+	recentRuns,
 	ownerName,
 	readOnly,
 	onToggleEnabled,
 	toggleDisabled,
 }: {
 	automation: SelectAutomation;
+	recentRuns: SelectAutomationRun[];
 	ownerName?: string | null;
 	readOnly?: boolean;
 	onToggleEnabled: (enabled: boolean) => void;
 	toggleDisabled?: boolean;
 }) {
+	const [tab, setTab] = useState<DetailTab>("settings");
 	const [name, setName] = useState(automation.name);
 	const [prompt, setPrompt] = useState(automation.prompt);
 	const lastSyncedPromptRef = useRef(automation.prompt);
@@ -34,8 +50,10 @@ export function AutomationBody({
 	}, [automation.prompt]);
 
 	const updateMutation = useMutation({
-		mutationFn: (patch: { name?: string }) =>
+		mutationFn: (patch: AutomationUpdatePatch) =>
 			apiTrpcClient.automation.update.mutate({ id: automation.id, ...patch }),
+		// The pickers re-render from the Electric-synced row, so a rejected
+		// update silently snaps back without this.
 		onError: (error) =>
 			toast.error(
 				error instanceof Error ? error.message : "Failed to update automation",
@@ -64,6 +82,16 @@ export function AutomationBody({
 		projectId: automation.v2ProjectId,
 	});
 
+	const { localHostId } = useWorkspaceHostOptions();
+	const hostId = automation.targetHostId ?? localHostId ?? null;
+	const hostUrl = useHostUrl(hostId);
+	const { agents: hostAgents } = useV2AgentChoices(hostUrl);
+	// Only warn once the host's terminal configs have loaded — the list always
+	// contains the built-in Superset chat entry, so length 1 means "not loaded
+	// yet / host unreachable", not "agent missing".
+	const agentMissing =
+		hostAgents.length > 1 && !matchAgentChoice(hostAgents, automation.agent);
+
 	return (
 		<div className="flex-1 overflow-y-auto px-8 py-8">
 			<div className="mx-auto flex w-full max-w-3xl flex-col">
@@ -81,7 +109,7 @@ export function AutomationBody({
 					placeholder="Automation title"
 					className="mb-3 text-2xl font-semibold"
 				/>
-				<div className="mb-8 flex items-center gap-2 text-sm">
+				<div className="flex items-center gap-2 text-sm">
 					<Switch
 						checked={automation.enabled}
 						onCheckedChange={onToggleEnabled}
@@ -100,22 +128,96 @@ export function AutomationBody({
 						</>
 					)}
 				</div>
-				<span className="mb-2 text-xs font-medium text-muted-foreground">
-					Instructions
-				</span>
-				<MarkdownEditor
-					content={prompt}
-					onChange={setPrompt}
-					editable={!readOnly}
-					onSave={(next) => {
-						if (readOnly) return;
-						if (next !== automation.prompt) {
-							setPromptMutation.mutate(next);
-						}
-					}}
-					placeholder="Add prompt e.g. look for crashes in $sentry"
-					searchFiles={searchFiles}
-				/>
+				{readOnly && (
+					<p className="select-text cursor-text mt-2 text-xs text-muted-foreground">
+						Owned by {ownerName ?? "a teammate"} — only they can edit this
+						automation.
+					</p>
+				)}
+
+				<div className="mt-6 mb-6 flex items-center gap-1">
+					{(
+						[
+							{ value: "settings", label: "Settings" },
+							{ value: "runs", label: "Run History" },
+						] as const
+					).map((t) => (
+						<button
+							key={t.value}
+							type="button"
+							onClick={() => setTab(t.value)}
+							className={cn(
+								"rounded-md px-3 py-1.5 text-sm transition-colors",
+								tab === t.value
+									? "bg-accent font-medium text-foreground"
+									: "text-muted-foreground hover:text-foreground",
+							)}
+						>
+							{t.label}
+						</button>
+					))}
+				</div>
+
+				{tab === "settings" ? (
+					<fieldset disabled={readOnly} className="contents">
+						<span className="mb-2 text-sm text-muted-foreground">Triggers</span>
+						<TriggersCard
+							automation={automation}
+							hostId={hostId}
+							readOnly={readOnly}
+							onUpdate={(patch) => updateMutation.mutate(patch)}
+						/>
+
+						<span className="mt-8 mb-2 text-sm text-muted-foreground">
+							Instructions
+						</span>
+						<div className="flex flex-col rounded-xl border border-border bg-card/40">
+							<div className="min-h-[240px] px-4 py-3">
+								<MarkdownEditor
+									content={prompt}
+									onChange={setPrompt}
+									editable={!readOnly}
+									onSave={(next) => {
+										if (readOnly) return;
+										if (next !== automation.prompt) {
+											setPromptMutation.mutate(next);
+										}
+									}}
+									placeholder="Add prompt e.g. look for crashes in $sentry"
+									searchFiles={searchFiles}
+								/>
+							</div>
+							<div className="flex items-center px-2.5 pb-2.5">
+								<AgentPicker
+									hostId={hostId}
+									disabled={readOnly}
+									value={automation.agent}
+									onChange={(id) => {
+										// The picker is scoped to `hostId` and emits a preset slug
+										// when unambiguous, falling back to the instance UUID. If
+										// the automation was previously auto-routed (targetHostId
+										// null), pin it to the host this value came from so a
+										// UUID-shaped agent can't be dispatched to a host that's
+										// never seen it.
+										const patch: AutomationUpdatePatch = { agent: id };
+										if (!automation.targetHostId && hostId) {
+											patch.targetHostId = hostId;
+										}
+										updateMutation.mutate(patch);
+									}}
+								/>
+							</div>
+						</div>
+						{agentMissing && (
+							<p className="select-text cursor-text mt-2 text-xs text-amber-600 dark:text-amber-500">
+								This agent no longer exists on the selected device (its agents
+								may have been reset). Runs will fail until you pick a new one.
+							</p>
+						)}
+					</fieldset>
+				) : (
+					<PreviousRunsList runs={recentRuns} />
+				)}
 			</div>
 		</div>
 	);
