--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/AutomationDetailSidebar.tsx
@@ -1,273 +1,0 @@
-import type {
-	SelectAutomation,
-	SelectAutomationRun,
-} from "@superset/db/schema";
-import {
-	formatDateTimeInTimezone,
-	nextOccurrenceAfter,
-} from "@superset/shared/rrule";
-import { toast } from "@superset/ui/sonner";
-import { Tooltip, TooltipContent, TooltipTrigger } from "@superset/ui/tooltip";
-import { useMutation } from "@tanstack/react-query";
-import { formatDistanceStrict } from "date-fns";
-import { useMemo } from "react";
-import { useRecentProjects } from "renderer/hooks/host-projects/useRecentProjects";
-import { useHostUrl } from "renderer/hooks/host-service/useHostTargetUrl";
-import { useV2AgentChoices } from "renderer/hooks/useV2AgentChoices";
-import { apiTrpcClient } from "renderer/lib/api-trpc-client";
-import { DevicePicker } from "renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/components/DevicePicker";
-import { useWorkspaceHostOptions } from "renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/components/DevicePicker/hooks/useWorkspaceHostOptions/useWorkspaceHostOptions";
-import { AgentPicker } from "../../../components/AgentPicker";
-import { ProjectPicker } from "../../../components/ProjectPicker";
-import { RelayOfflineNotice } from "../../../components/RelayOfflineNotice";
-import { SchedulePicker } from "../../../components/SchedulePicker";
-import { TimezonePicker } from "../../../components/TimezonePicker";
-import { WorkspacePicker } from "../../../components/WorkspacePicker";
-import { matchAgentChoice } from "../../../utils/agentIdentity";
-import { PreviousRunsList } from "../PreviousRunsList";
-import { Row } from "./components/Row";
-import { Section } from "./components/Section";
-import { SectionTitle } from "./components/SectionTitle";
-
-interface AutomationDetailSidebarProps {
-	automation: SelectAutomation;
-	recentRuns: SelectAutomationRun[];
-	ownerName?: string | null;
-	/** Disables every editor — automation updates are owner-gated server-side. */
-	readOnly?: boolean;
-}
-
-export function AutomationDetailSidebar({
-	automation,
-	recentRuns,
-	ownerName,
-	readOnly,
-}: AutomationDetailSidebarProps) {
-	const recentProjects = useRecentProjects();
-	const { localHostId } = useWorkspaceHostOptions();
-	const selectedProject = recentProjects.find(
-		(p) => p.id === automation.v2ProjectId,
-	);
-
-	const hostId = automation.targetHostId ?? localHostId ?? null;
-
-	const hostUrl = useHostUrl(hostId);
-	const { agents: hostAgents } = useV2AgentChoices(hostUrl);
-	// Only warn once the host's terminal configs have loaded — the list always
-	// contains the built-in Superset chat entry, so length 1 means "not loaded
-	// yet / host unreachable", not "agent missing".
-	const agentMissing =
-		hostAgents.length > 1 && !matchAgentChoice(hostAgents, automation.agent);
-
-	const updateMutation = useMutation({
-		mutationFn: (
-			patch: Partial<
-				Parameters<typeof apiTrpcClient.automation.update.mutate>[0]
-			>,
-		) =>
-			apiTrpcClient.automation.update.mutate({ id: automation.id, ...patch }),
-		// The pickers re-render from the Electric-synced row, so a rejected
-		// update silently snaps back without this.
-		onError: (error) =>
-			toast.error(
-				error instanceof Error ? error.message : "Failed to update automation",
-			),
-	});
-
-	const lastRunAt = recentRuns
-		.map((run) => run.scheduledFor)
-		.map((d) => (d ? new Date(d) : null))
-		.filter((d): d is Date => d !== null)
-		.sort((a, b) => b.getTime() - a.getTime())[0];
-
-	// Paused automations keep a stale nextRunAt; compute what the schedule
-	// would fire next so edits are previewable before resuming.
-	const nextRunDate = useMemo(() => {
-		if (automation.enabled) {
-			return automation.nextRunAt ? new Date(automation.nextRunAt) : null;
-		}
-		try {
-			return nextOccurrenceAfter({
-				rrule: automation.rrule,
-				dtstart: new Date(automation.dtstart),
-				timezone: automation.timezone,
-				after: new Date(),
-			});
-		} catch {
-			return null;
-		}
-	}, [
-		automation.enabled,
-		automation.nextRunAt,
-		automation.rrule,
-		automation.dtstart,
-		automation.timezone,
-	]);
-
-	return (
-		<aside className="flex w-[360px] shrink-0 flex-col overflow-hidden border-l border-border">
-			<div className="flex shrink-0 flex-col gap-6 px-5 pt-5 pb-2">
-				<Section title="Status">
-					<Row
-						label="Next run"
-						value={
-							nextRunDate ? (
-								<Tooltip>
-									<TooltipTrigger asChild>
-										<span>
-											{automation.enabled
-												? formatDistanceStrict(nextRunDate, new Date(), {
-														addSuffix: true,
-													})
-												: `Would run ${formatDistanceStrict(nextRunDate, new Date(), { addSuffix: true })}`}
-										</span>
-									</TooltipTrigger>
-									<TooltipContent side="left">
-										{formatDateTimeInTimezone(nextRunDate, automation.timezone)}
-									</TooltipContent>
-								</Tooltip>
-							) : (
-								"—"
-							)
-						}
-					/>
-					<Row
-						label="Last ran"
-						value={
-							lastRunAt
-								? formatDateTimeInTimezone(lastRunAt, automation.timezone)
-								: "—"
-						}
-					/>
-				</Section>
-
-				<Section title="Details">
-					{readOnly && (
-						<p className="select-text cursor-text pb-2 text-xs text-muted-foreground">
-							Owned by {ownerName ?? "a teammate"} — only they can edit this
-							automation.
-						</p>
-					)}
-					{/* display:contents keeps the rows in the section's flex column;
-					    disabled propagates to every picker's trigger button. */}
-					<fieldset disabled={readOnly} className="contents">
-						<Row
-							label="Device"
-							value={
-								<DevicePicker
-									className="-mr-4"
-									hostId={hostId}
-									showLocalOnlineState
-									disabled={readOnly}
-									onSelectHostId={(nextHostId) => {
-										updateMutation.mutate({ targetHostId: nextHostId });
-									}}
-								/>
-							}
-						/>
-						<Row
-							label="Project"
-							value={
-								<ProjectPicker
-									className="-mr-4"
-									selectedProject={selectedProject}
-									sessionSelected={automation.v2ProjectId === null}
-									recentProjects={recentProjects}
-									onSelectProject={(v2ProjectId) =>
-										updateMutation.mutate({ v2ProjectId })
-									}
-								/>
-							}
-						/>
-						<Row
-							label="Workspace"
-							value={
-								<WorkspacePicker
-									className="-mr-4"
-									hostId={hostId}
-									projectId={automation.v2ProjectId}
-									value={automation.v2WorkspaceId}
-									onChange={(v2WorkspaceId) =>
-										updateMutation.mutate({
-											v2WorkspaceId,
-											// Denormalized pin: the picker is scoped to this
-											// host/project, so send both — the cloud stores them
-											// without a workspace-registry lookup. A null project
-											// means the pin is a session workspace.
-											...(v2WorkspaceId && hostId
-												? {
-														targetHostId: hostId,
-														v2ProjectId: automation.v2ProjectId,
-													}
-												: {}),
-										})
-									}
-								/>
-							}
-						/>
-						<Row
-							label="Repeats"
-							value={
-								<SchedulePicker
-									className="-mr-4"
-									rrule={automation.rrule}
-									onRruleChange={(rrule) => updateMutation.mutate({ rrule })}
-								/>
-							}
-						/>
-						<Row
-							label="Agent"
-							value={
-								<AgentPicker
-									className="-mr-4"
-									hostId={hostId}
-									disabled={readOnly}
-									value={automation.agent}
-									onChange={(id) => {
-										// The picker is scoped to `hostId` and emits a preset slug
-										// when unambiguous, falling back to the instance UUID. If
-										// the automation was previously auto-routed (targetHostId
-										// null), pin it to the host this value came from so a
-										// UUID-shaped agent can't be dispatched to a host that's
-										// never seen it.
-										const patch: { agent: string; targetHostId?: string } = {
-											agent: id,
-										};
-										if (!automation.targetHostId && hostId) {
-											patch.targetHostId = hostId;
-										}
-										updateMutation.mutate(patch);
-									}}
-								/>
-							}
-						/>
-						{agentMissing && (
-							<p className="select-text cursor-text text-xs text-amber-600 dark:text-amber-500">
-								This agent no longer exists on the selected device (its agents
-								may have been reset). Runs will fail until you pick a new one.
-							</p>
-						)}
-						<Row
-							label="Timezone"
-							value={
-								<TimezonePicker
-									className="-mr-4"
-									value={automation.timezone}
-									onChange={(timezone) => updateMutation.mutate({ timezone })}
-								/>
-							}
-						/>
-					</fieldset>
-					<RelayOfflineNotice hostId={hostId} className="mt-2" />
-				</Section>
-			</div>
-
-			<div className="mt-6 flex min-h-0 flex-1 flex-col gap-2 pl-5 pr-3 pb-5">
-				<SectionTitle>Previous runs</SectionTitle>
-				<div className="min-h-0 flex-1 overflow-y-auto">
-					<PreviousRunsList runs={recentRuns} />
-				</div>
-			</div>
-		</aside>
-	);
-}
