--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/components/Row/Row.tsx
@@ -1,10 +1,0 @@
-import type { ReactNode } from "react";
-
-export function Row({ label, value }: { label: string; value: ReactNode }) {
-	return (
-		<div className="flex min-h-8 items-center gap-4 text-sm">
-			<span className="shrink-0 text-muted-foreground">{label}</span>
-			<div className="flex min-w-0 flex-1 justify-end">{value}</div>
-		</div>
-	);
-}
