--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/components/Section/Section.tsx
@@ -1,17 +1,0 @@
-import type { ReactNode } from "react";
-import { SectionTitle } from "../SectionTitle";
-
-export function Section({
-	title,
-	children,
-}: {
-	title: string;
-	children: ReactNode;
-}) {
-	return (
-		<section className="flex flex-col gap-3">
-			<SectionTitle>{title}</SectionTitle>
-			<div className="flex flex-col">{children}</div>
-		</section>
-	);
-}
