--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/components/Section/index.ts
@@ -1,1 +1,0 @@
-export { Section } from "./Section";
