--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/components/SectionTitle/SectionTitle.tsx
@@ -1,9 +1,0 @@
-import type { ReactNode } from "react";
-
-export function SectionTitle({ children }: { children: ReactNode }) {
-	return (
-		<span className="font-sans text-xs font-medium uppercase tracking-wider text-muted-foreground">
-			{children}
-		</span>
-	);
-}
