--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/components/SectionTitle/index.ts
@@ -1,1 +1,0 @@
-export { SectionTitle } from "./SectionTitle";
