--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/AutomationDetailSidebar/index.ts
@@ -1,1 +1,0 @@
-export { AutomationDetailSidebar } from "./AutomationDetailSidebar";
