+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/$automationId/components/TriggersCard/index.ts
@@ -1,0 +1,1 @@
+export { type AutomationUpdatePatch, TriggersCard } from "./TriggersCard";
